+++ components/blockissuer/types.go
@@ -5,5 +5,5 @@ type BlockIssuerInfo struct {
 	// The account address of the block issuer.
 	BlockIssuerAddress string `json:"blockIssuerAddress"`
 	// The number of trailing zeroes required for the proof of work to be valid.
-	PowTargetTrailingZeros int `json:"powTargetTrailingZeros"`
+	PowTargetTrailingZeros uint8 `json:"powTargetTrailingZeros"`
 }
